refactor package to work with hypercorn
