+++ .gitignore
@@ -145,3 +145,4 @@ dmypy.json
 cython_debug/
 
 # PyCharm
+poetry.lock
