+++ src/acrossword/__init__.py
@@ -1,4 +1,4 @@
 from .documents.documents import Document, DocumentCollection
 from .rankers.rank import Ranker
-from .rankers.server import run
+from .server import run
 from .rankers.classifier import Classifier, Category
